fix: switch frontend vite config to esm for docker build
