+++ package.json
@@ -22,9 +22,9 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "docker": "docker compose down -v && docker compose up --build",
-    "frontend:dev": "vite dev --config frontend/vite.config.ts --host 0.0.0.0 --port 3055",
-    "frontend:build": "vite build --config frontend/vite.config.ts",
-    "frontend:start": "vite preview --config frontend/vite.config.ts --host 0.0.0.0 --port 3055",
+    "frontend:dev": "vite dev --config frontend/vite.config.mjs --host 0.0.0.0 --port 3055",
+    "frontend:build": "vite build --config frontend/vite.config.mjs",
+    "frontend:start": "vite preview --config frontend/vite.config.mjs --host 0.0.0.0 --port 3055",
     "dev:full": "npm run dev",
     "dev": "concurrently \"npm run start:dev\" \"npm run frontend:dev\""
   },
